requirements.txt
asgiref==3.6.0
autopep8==2.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.4
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-import-export==3.0.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pytz==2022.7
PyYAML==6.0
sqlparse==0.4.3
tablib==3.3.0
tomli==2.0.1
tzdata==2022.7
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
